backend/requirements.txt
Django
djangorestframework
django-cors-headers
numpy
kokoro
openai-whisper
torch
torchaudio
